+++ src/activities/boot_sleep/SleepActivity.h
@@ -22,6 +22,7 @@ class SleepActivity final : public Activity {
   void renderCoverSleepScreen() const;
   void renderReadingStatsSleepScreen() const;
   void renderMinimalSleepScreen() const;
+  void renderMinimalStatsSleepScreen() const;
   void renderBitmapSleepScreen(const Bitmap& bitmap) const;
   void renderLastScreenSleepScreen() const;
   void renderBlankSleepScreen() const;